move navbar into its own dir
